Use sparsehash/dense_hash_{set,map}
We have quite a few choices for set/map structures now; at some point,
we should work out a common interface and turn users into templates.  By
calling setrlimit we should be able to get bad_alloc instead of
swapping/OOM-killing, so we can try the fast-but-memory-hungry
implementation and fall back to slow-but-space-efficient
implementations.