Disable intersection folding; intersect column constraints with puzzle size
Folding e.g., concats is always going to be faster due to removing some
extra copies, but intersection's quadratic behavior means we might
prefer to do two small intersections and one large instead of one huge
intersection, and folding prevents client code from controlling that.

We don't need to use the size constraint for its own sake, because each
row in the row constraint has a width constraint, and the size
constraint is just the concatenation of those width constraints.

Intersecting the size constraint with the column constraints helps
because their height is not otherwise constrained.