Nonogram: upper bound on zeroes
The idea is that intersections will have fewer dead-ends because they
can't get stuck in star/plus loops.  This isn't a clear win, because
automata with upper bounds are larger due to the counting.  In practice
it seems like a mild improvement on average but doesn't help the
worst-case/peak state size.