PackedAutomaton and an implementation
These rely on forming a pointer past the end of an object to access
coallocated data.  That's technically undefined behavior, but it seems
to work, and I don't know of a conforming way that doesn't involve
storing a pointer in PackedAutomaton that points directly after the
object, and then casting that pointer to access the data anyway.