Replace std::stack with circular_deque (other half of issue #1)
We could instead use std::vector (or small_vector).  It might be
able to memcpy trivial types more often because it's always linear.

We could use larger small allocations, because we aren't recursive
and thus aren't too worried about stack usage.

We could replace Tarjan's use of std::vector with circular_deque.
Tarjan uses std::find to check containment, and I'm worried that
won't be a memcmp with std::deque because begin() and end() have to
cope with nonlinearity (even through we aren't using it because
we're always a stack).