finish must canonicalize with respect to rotations
Previously we were calling the standard, doesn't-change-the-language
canoncalizer from finish.  But that's clearly not right, because we do
change the language when canonicalizing known gadgets and mirror's return
value.

With this fix, the generational search now does C2T -> A2T and C2T -> P2T
in two generations, as expected.