mirror is canonicalization with mirrored permutations only (fixes #9)
This goes back to the "canoncalizeRenumber is a composition" test cases.
I also empirically tested on automata from the generational search: both
the old and new implementations built the same automata.

This is faster, but mirror's already pretty fast anyway.  The actual
payoff is clarifying when we need to mirror during the search.  We
definitely don't need to mirror before connect now, because the mirrored
connect outputs will just get canonicalized back when finishing.