Use edge list through determinize and removeDeadStates (part of #3)
Currently we return to the usual representation before hopcroft, but that
will change in a future commit.

So far this is moderately slower, mostly because we're sorting so much.
We can investigate using another sort algorithm, not using the
Edge::forwards/backwards function pointers, building an index in
removeDeadStates like the old one does, and not bothering to sort for
locality.