AutomatonBase::run simplification (#14)
I deleted the iterator pair and iterator range overloads as they aren't
used.  If we actually need them, we can easily have the initializer list
overload delegate to a (nontemplated) pointer pair overload, which is
good enough for vectors, the most likely calling type.

Using numeric_cast instead of static_cast costs only a tiny amount and
will give a better error message if we accidentally pass a negative
number here, so it's worth it.