TBB-parallelize connect
The functional form of tbb:parallel_reduce is forced to make lots of
copies, but we do get speedup.  We may want to sort the SCCs by size,
then shuffle or otherwise load-balance the non-singleton components,
because if all the large components are in the left division, work
stealing won't be that effective (can't pre-empt once we start
evaluating a range).

Also I think we'll have a dedicated TBB-parallelized search form, at
least for now, so we'll use Finisher directly as our reduction return
value.  Or maybe a vector or rope for individual connects at a specific
location, and a Finisher at the connect level and above.