Remove the queue-based toggles search and toggles benchmark
The queue-based search hasn't been maintained for a while, so while it
compiles and runs, it's no longer a useful point of comparison.

The toggles benchmark needs to be updated now that we're using TBB.

Mostly I'm trying to remove clients of the old PackedAutomaton system so
I can remove it, too.