Update benchmark-pack to new pack scheme
As a bonus, we can now easily print the total size of the packs.  We
might want to print the size of each pack in some diff-friendly format
so we can track the benefit of new coding flags.  (The format for binary
size diffs, used in earlier commit messages, would work well.)