Don't mirror on the left
I thought that if we failed to mirror on the right we had to try
mirroring on the left, but empirically it isn't the case.  We could also
mirror on the left and not on the right, but obviously we'd prefer the
current state because this way we only have to do the actual mirroring
once.