Epsilon closure experiment (doesn't help)
The idea here is that we're adding nondeterminism with calls to
addEpsilon, so we should try to remove it.  Right-optimizing already
removes the nondeterminism that we'd get by merging connected components
in the epsilon-edge graph (though maybe we could do it faster looking at
just that graph).  So then I tried taking the closure of the epsilon
graph, which would reduce the number of DFA states generated in
determinize at the cost of increasing the size of each NFA state set.
But it turns out that we don't add any edges in the process, so we can't
get any help there.