Determinize uses smaller NFA state type when possible (close #43)
A small but measurable improvement in max resident size with no cost in
runtime.

We can't use a smaller type for DFA states because we don't know in
advance how many there will be, and because we're unlikely to save any
space due to alignment issues in the hash table.