Loopback splitting experiment (makes things worse)
I noticed in the optimized automata we somewhat frequently had nonaccept
states that differed only in the presence of a loopback edge.  This code
creates a new state for the loopback edge, then reuses a state without a
loopback edge.  The accept state then has two transitions on that
symbol, one to the loopback state and the other to the state
representing "actual" gadget traversal.

Unfortunately this ends up increasing the blowup.