Make WorkerManager::run do something on the second call
io_context::restart() doesn't seem to be mentioned anywhere.  I thought
for a while I needed to destroy and reconstruct the socket objects, but
that doesn't seem to be necessary.

We also clear the buffers before calling generator->next() in run() and
reconnect if there's still work to do after reading.