Remove completed_combines table
It doesn't make sense to store a precision value for each pair of
gadgets, because either all of their combinations fit in a given
precision or none of them do.  (Rotating and moving the splice point
don't affect the alphabet size of the intermediate automaton.)

It doesn't make sense to have a separate table of completion
information.  A combine between two gadgets has been done if there are
any combine edges with those gadgets as inputs, and hasn't been done
otherwise.  (It's possible the combine resulted in a single vertex after
connecting and we don't store those, but those should be cheap enough to
recompute that it's not worth a table just for them.)

Completion tables are useful for unary operations because they can use
ranges (the table is an interval set), but that's less easily
implementable for binary operations.  We could have a table with input2
as a scalar column and input1 a range column, and check if a left
input is in any of the ranges for a particular right input's rows,
should checking the edge table become too expensive.