toggles-runner: send summary stats, not gadget ids
The driver has to follow edges from the database after each operation
because some of the computations might have been performed on a previous
run, so the driver can't use the gadget ids the runner was sending.  The
driver does want to compute some basic summary statistics, though, so
we'll send those instead.

The driver is still ignoring responses (not even checking for errors).