Fix GadgetBuilder initial state handling
First, when NDEBUG is defined, we skipped the asserts, so the first
addState in the gadget was the half-gadget-edge intermediate state, not
a state corresponding to a gadget state.  That meant the gadget's
language was always wrong (didn't accept the empty state) and usually
had the wrong number of states and/or locations.

But looking at GadgetBuilder's callers, it's clear I misunderstood what
the preallocated_states parameter meant.  It's the number of gadget
states to reserve, not the automaton's states.  So now GadgetBuilder
users just pass the number of gadget states, to be sure.