Print things we find
We need to close (and mirror?) the targets.  We probably want to remove
found targets (and their close/mirror ancestors?) so we don't print them
every generation, and so we know when we can stop following edges.  We
need to remember human-friendly names (i.e., what was passed on the
command line), or look them up.