Reports include source gadget names
I considered trying to give names for every gadget, and that might still
be useful, but whether a name gets used depends on the hierarchical
decomposition of the combined gadget.  For example, two connected
3-splits are a 4-split, but if we connected a 3-split to something else
already, we won't recognize the 4-split.

That raises the idea of discovering additional edges without computing
them based on the hierarchy, but I don't see a clear path to doing that
now.