Driver prepares combine statements just before use
Now that we can resume from checkpoint, there's no single place to do
these prepares.  You might think we could do it in the
Search(Serialized&&) constructor, but we might have suspended before
filling out combine_rights.  It's easiest to make it idempotent and do
it before we need it.