Parallelize get_combines
About the same 2x speedup.

This does change the claimed number of gadgets at the end of edges, but
that's just because the database only deduplicates (via group by) within
queries and we're now running a different set of queries.  The number
after deduplication in the SearchState is the same.