Gadgets and names sync working
This depends on some changes to the lmdbxx wrapper (in vendor/).
lmdb::to_sv just seems broken: it returns a pointer to the function's
parameter, which is destroyed at the end of the function.

This commit moved a bunch of code to avoid having to forward-declare
selsert_gadget_by_data, but some of that code changed, too.  Sorry about
that.