Introduce max left states combine skip condition
Gadgets with lots of states often have memory-hungry determinizes.  We
can defer doing those combines until we've searched further in the
easy-to-search space.  We write completions based on provenance, so we
can't accidentally record that we're done for the combines we skip.

The driver was slightly refactored in anticipation of adding more such
completeness-affecting options.

As a side effect, this implements #49 (skipping before decoding) for
combine.  We still copy all the data in select_gadget_id_to_data; if we
want to filter earlier, this change makes it easier by letting us just
leave things out of the left_data vector instead of messing with the
map and intervals.

We might want state-based predicates to help the driver (relates #50).