Tunnels definition
Transcribed from
https://coauthor.csail.mit.edu/6.892-2019/m/7aTtAZckguMm74Yhy , probably
with some errors.  This file isn't intended to be sync'd directly, but
instead as input for systematic generation of all pairs of tunnels, plus
the tunnels that make sense on their own.

These names could be improved for our purposes, because they don't
compose well into pairs.  Also the "locked toggle" isn't related to
either the locking 2-toggles or the toggle-locks.