Driver can launch its own socat process
In theory, using sockets means we can run runners on other machines.  In
practice, we're now running all runners on the same machine so they can
work directly on the database, and if we wanted to farm work out to more
nodes, we'd revive the batch task file writing/reading code instead of
running live.  Given that, it's very convenient for the driver to launch
its own workers.

Boost.Process claims it works with asio but it isn't obvious how and
there are some GitHub issues suggesting it doesn't work correctly for
multiple processes, so we'll do the easy thing and launch a socat
process just like the human would do manually.