Attempt at gadget index migration (#91)
This is complicated and not actually as transactional as I intended it
to be.  I think it'll be easier to just dump all the hashes to a temp
file (this is sequential write, so can be on spinning rust), map that
file, drop the database and append pages from the mapped data.

Still, here it is in case I want to come back to it.