Fix two selsert corner cases
If we pruned everything early, we didn't initialize novel_global_ids.

If we didn't prune everything early but pruned the rest late, we'd try
to add zero hashes to the current hash page if it isn't full.  This
would throw due to MDB_NOOVERWRITE (assuming that check isn't optimized
away).