Fix (?) gadget encoding components weirdness (see #92)
I don't know what was actually wrong here, or why it goes away when we
assign automaton::find_components(a) to a local variable.  But here's
hoping this actually fixes it.

Note that we don't compute reachability anymore.  The automaton is
minimal, so it doesn't have unreachable components.  (Back when we were
using superposition, we wanted to do this before minimizing, because
we were going to put all the accept states in a component in
superposition and minimize again.)