Fix memory retention in reporter's merge_unique_vectors
Before, the recent transform_reduce change (#102) increased memory
consumption; after, it decreases it as expected.

C++ is frustrating in this regard: if we take parameters by rvalue
reference, we don't actually release the memory, but if we take them by
value, it's easy to accidentally make a copy by not moving.