Remove nop-only locations (#88)
In a short test running toggle-toggle-parallel and 3-split for 4
generations with precision 10, we find 152496 gadgets with this change
compared to 183333 without, while finding the same 33 named gadgets
(some via different traces).

The code could be improved to find the active alphabet in the same
iteration (not allowing for breaking the loop early).  I should also
type in comments from the cell-phone photograph of a piece of paper
where I worked out why finding these small loops in the automaton is
sufficient.