Move Automaton regexesque functions to separate header (#119)
automaton-regexesque-basic.hpp is necessary because empty/epsilon/all
are convenient things to assign to *this and I didn't want to inline
them multiple times.

same_language/compare_languages use use comp and conj, so are also
moved.  These are mostly used in tests.  Comparing after canonicalizing
is likely to be more performant, anyway.