Nagle-like algorithm for secondhalf launch
This helps somewhat, but predicates will be larger under the
commit-by-slice system.  If this is a problem, we could teach the
runner's writer thread to commit multiple slices at once when slices are
small.

While we're in the area, we also now pass the number of secondhalf
reader threads from the driver instead of hard-coding it.  Too few
threads and the writer thread won't be fully utilized; too many threads
and we'll lose the prefetching benefit.  If we can't find an acceptable
number, we could teach the runner's reader threads to cooperate on a
single slice instead of one slice per thread.