Nice firsthalf worker processes
We want the secondhalf task (if there is one) to always be scheduled,
but we don't want to dedicate threads to it at other times.  This change
should allow using more secondhalf reader threads without reserving
them when no secondhalf is running.

There seems to be no performance penalty for setting the maximum
niceness, so we do that by default.