Fix deleted locations reporting (broken in f6ed78b)
f6ed78b changed connect_at to not return the active alphabet, which is
good because we only need it for reporting deleted locations.  But
connect_at copies its input automaton instead of modifying it in place,
and the adjusted code in connect_deleted_symbols checked the input
automaton, so we never saw any deleted locations.

We now return a bitset, which should be cheap, and decode it only when
we're reporting.  This shouldn't impact actual connects.

This caused some reports on archived databases to be wrong.