Delete driver checkpoint code
This was never that useful.  We partially ported it to LMDB, but
resuming the hard way has proved to be fast enough.

The distinction between completeness and runtime options is useful
documentation, so we'll preserve that, and some comments were left in
case we want to reimplement checkpointing in the future.