Beginnings of SkinnyProv compression (#104)
This doesn't actually compress anything yet, and so has higher peak
memory consumption due to extra copying.  It's also a lot less smart
than I was planning, but if we're willing to pack 8 MiB of provs into
each block, we'll only have ~10000 blocks, so for < 1000 queries it's
not too expensive to just check all the blocks.  (That means the two
sorted block lists aren't being taken advantage of and will be removed.)