Reporter: accumulate provs in deque to reduce peak memory
We use boost::container::deque because it lets us specify a large chunk
size (using 1 MiB, compared to 512 bytes of libstdc++).  I checked the
code and it doesn't seem to have a clownshoes problem.

We probably also want to use deques for our pending interval lists, if
not compress them entirely, also to reduce peak consumption.