Canonicalize using maximal nop set (#103)
We now add the maximal set of nop edges to the automaton before
canonicalizing, remove them when encoding, and restore them when
decoding the automaton (because GadgetBuilder::build canonicalizes).
This seems to dramatically reduce the number of gadgets found, because
gadgets can no longer be distinguished only by nop edges.

We still remove nop-only locations.  I think we still have to do this
before encoding for canonicalization to be correct.

It shouldn't be necessary to add the nop edges when decoding, and not
doing so might speed up our automata operations.