Add script to generate input/output gadgets
We also generate fanin/fanout and mixer gadgets here, because the usual
model of input/output gadgets includes fanin (but not fanout) and the
mixers are convenient to generate with the bracelet-calculating code.

This is the first client of the JSON sync parsing.  Going forward, only
files intended to be manually edited by humans will use YAML.